chore(nx-dev): switch to playwright
