+++ docs/nx-cloud/private/auth-saml.md
@@ -134,7 +134,7 @@ if you are interested.
     1. `SAML_CERT=<your-cert-string-from-above>`
     2. `SAML_ENTRY_POINT=<your-login-url-from-above>`
 
-# Helm config
+## Helm config
 
 If you are using [Helm to deploy NxCloud](https://github.com/nrwl/nx-cloud-helm) you
 will need to configure the below two values, as well as make the `SAML_CERT` and `SAML_ENTRY_POINT`
