+++ nx-dev/nx-dev-e2e/.eslintrc.json
@@ -1,9 +1,9 @@
 {
-  "extends": ["plugin:cypress/recommended", "../../.eslintrc.json"],
+  "extends": ["plugin:playwright/recommended", "../../.eslintrc.json"],
   "ignorePatterns": ["!**/*"],
   "overrides": [
     {
-      "files": ["*.ts", "*.tsx", "*.js", "*.jsx"],
+      "files": ["src/**/*.{ts,js,tsx,jsx}"],
       "rules": {}
     }
   ]
