--- nx-dev/nx-dev-e2e/src/e2e/app.cy.ts
@@ -1,10 +1,0 @@
-describe('nx-dev', () => {
-  beforeEach(() => cy.visit('/'));
-
-  it('should display the primary heading', () => {
-    cy.get('[data-cy="primary-heading"]').should(
-      'contain',
-      'Smart, FastExtensibleBuild System'
-    );
-  });
-});
