+++ nx-dev/nx-dev/project.json
@@ -22,9 +22,8 @@
     },
     "sitemap": {
       "executor": "nx:run-commands",
-      "outputs": ["{options.outputPath}"],
+      "outputs": ["{workspaceRoot}/dist/nx-dev/nx-dev/public"],
       "options": {
-        "outputPath": "dist/nx-dev/nx-dev/public",
         "command": "pnpm next-sitemap --config ./nx-dev/nx-dev/next-sitemap.config.js"
       }
     },
