+++ package.json
@@ -72,6 +72,7 @@
     "@nx/js": "16.7.0-rc.1",
     "@nx/linter": "16.7.0-rc.1",
     "@nx/next": "16.7.0-rc.1",
+    "@nx/playwright": "16.7.0-rc.1",
     "@nx/react": "16.7.0-rc.1",
     "@nx/storybook": "16.7.0-rc.1",
     "@nx/web": "16.7.0-rc.1",
@@ -167,6 +168,7 @@
     "eslint-plugin-cypress": "2.14.0",
     "eslint-plugin-import": "2.26.0",
     "eslint-plugin-jsx-a11y": "6.6.1",
+    "eslint-plugin-playwright": "^0.15.3",
     "eslint-plugin-react": "7.31.11",
     "eslint-plugin-react-hooks": "4.6.0",
     "eslint-plugin-storybook": "^0.6.12",
